### 2018-06-11 25b41eb
`proto_event` now handles all event calculations 'in house', allowing `detect_event` to now be given a theoretically limitless number of thresholds
