+++ R/proto_event.R
@@ -10,67 +10,93 @@
 #' from \code{\link{detect_event}} within which this runs.
 #' @param criterion_column The column to use for the detection of events.
 #' @param minDuration Minimum duration for acceptance of detected events.
-#' @param gaps This logic gate tells this internal function if it should be
-#' calculating the lengths of events, or rather the gaps in between them.
-#' The default is \code{FALSE}.
+#' @param joinAcrossGaps This logic gate tells this internal function if
+#' it should connect events across the \code{maxGap} (see below). The
+#' default it inherits is \code{TRUE}.
 #' @param maxGap This is the number of rows (days) across which distinct
 #' events will be combined into one event if \code{joinAcrossGaps = TRUE}.
 #'
 #' @return A dataframe that will be used within
 #' \code{\link{detect_event}}.
 #'
-#' @author Smit, A. J.
+#' @author Albertus J. Smit, Robert W. Schlegel
 #'
 proto_event <- function(t_series,
                         criterion_column,
                         minDuration,
-                        gaps = FALSE,
+                        joinAcrossGaps,
                         maxGap) {
 
   index_start <- index_end <- duration <- NULL
 
-  ex <- rle(as.vector(eval(substitute(criterion_column), t_series)))
-  # ex <- rle(as.vector(t_series[, get(criterion_column)]))
-  ind <- rep(seq_along(ex$lengths), ex$lengths)
-  s <- split(1:nrow(t_series), ind)
+  ex1 <- rle(criterion_column)
+  ind1 <- rep(seq_along(ex1$lengths), ex1$lengths)
+  s1 <- split(1:nrow(t_series), ind1)
 
-  if (!(gaps)) {
+  proto_events <- do.call(rbind,
+                          lapply(s1[ex1$values == TRUE], function(x)
+                            data.table::data.table(index_start = min(x), index_end = max(x))))
+  proto_events$duration <- proto_events$index_end - proto_events$index_start + 1
+  proto_events <- proto_events[duration >= minDuration, ]
 
-    proto_event_value <- s[ex$values == TRUE]
+  if (length(proto_events$index_start) == 0)
+    stop("not enough consecutive days above the threshold(s) to detect any events")
 
-    proto_event_rng <-
-      lapply(proto_event_value, function(x)
-        data.table::data.table(index_start = min(x), index_end = max(x)))
+  durationCriterion <- rep(FALSE, nrow(t_series))
+  for (i in 1:nrow(proto_events)) {
+    durationCriterion[proto_events$index_start[i]:proto_events$index_end[i]] <-
+      rep(TRUE, length = proto_events$duration[i])
+  }
 
-    proto_events <- do.call(rbind, proto_event_rng)
-    proto_events$event_no <- cumsum(ex$values[ex$values == TRUE])
-    proto_events$duration = proto_events$index_end - proto_events$index_start + 1
-    proto_events <- proto_events[duration >= minDuration, ]
-    proto_events$date_start <- t_series$ts_x[proto_events$index_start]
-    proto_events$date_end <- t_series$ts_x[proto_events$index_end]
+  if (joinAcrossGaps) {
+    ex2 <- rle(durationCriterion)
+    ind2 <- rep(seq_along(ex2$lengths), ex2$lengths)
+    s2 <- split(1:nrow(t_series), ind2)
 
-    return(proto_events)
-
-  } else {
-
-    proto_gap_value <- s[ex$values == FALSE]
-
-    proto_gap_rng <-
-      lapply(proto_gap_value, function(x)
[…]
