### 2018-06-18 #768 @robwschlegel
Hello,
For some reason this e-mail only came through to my inbox now.
There were a couple of issues with the original manuscript that I have
attended to. The manuscript on the JOSS site therefore needs to be
re-compiled.
All the best,
-Robert
