### 2018-06-18 cb4a7e6
Updates to paper manuscript
