### 2018-06-22 e290323
Re-submitted to CRAN in anticipation of __`ggplot2`__ changes
