### 2018-06-25 #768 @leeper
@dhhagan @geanders Just another quick ping - would either of you be interested in reviewing this for JOSS?
