### 2018-06-25 #768 @leeper
@adamhsparks Would you be willing to review this package for JOSS?
