### 2018-06-25 #768 @geanders
I'm in the same boat. It looks like a really interesting package, but it would take me a couple of weeks to get to it.
