### 2018-06-25 #768 @adamhsparks
Sorry, mate, I'm singing from the same hymnal. I have three conferences in
July/August. I hate to say no, but have to.

On Tue, 26 Jun 2018 at 03:13, Thomas J. Leeper <notifications@github.com>
wrote:

> @geanders <https://github.com/geanders> No problem. May come back to you.
> Thanks!
>
