### 2018-07-10 7c65741
Bug fix. Up to v0.3.1.
