+++ src/clim_calc.cpp
@@ -28,9 +28,9 @@ Rcpp::NumericMatrix clim_calc_cpp(arma::mat x, int windowHalfWidth, double pctil
   std::iota(id.begin(), id.end(), 0);
 
   // allocate receiving vectors
-  arma::vec s_raw(ydim);
-  arma::vec t_raw(ydim);
-  arma::vec v_raw(ydim);
+  arma::vec s_raw(ydim); // climatology
+  arma::vec t_raw(ydim); // threshold
+  arma::vec v_raw(ydim); // variance
 
   for(int i = windowHalfWidth; i < ydim - windowHalfWidth; i++) {
     arma::mat window = x.submat(i - windowHalfWidth, 0, i + windowHalfWidth, xdim - 1);
@@ -40,7 +40,7 @@ Rcpp::NumericMatrix clim_calc_cpp(arma::mat x, int windowHalfWidth, double pctil
     v_raw[i] = stddev(v);
   }
 
-  //trim tops and bottoms (return 366 day climatology)
+  // trim tops and bottoms (return 366 day climatology)
   arma::vec doy(366);
   std::iota(doy.begin(), doy.end(), 1);
   arma::vec seas = s_raw.subvec(windowHalfWidth, ydim - windowHalfWidth - 1);
