### 2018-07-11 #768 @leeper
@khaors Would you be willing to review this software for JOSS?
