### 2018-07-12 #821 @whedon
[REVIEW]: heatwaveR: A central algorithm for the detection of heatwaves and cold-spells
**Submitting author:** @robwschlegel (<a href="http://orcid.org/0000-0002-0705-1287">Robert Schlegel</a>)
**Repository:** <a href="https://github.com/robwschlegel/heatwaveR" target ="_blank">https://github.com/robwschlegel/heatwaveR</a>
**Version:** 0.2.7
