### 2018-07-12 a6aec6c
Code streamlining. Up to v0.3.2.
