### 2018-07-12 #821 @leeper
@khaors The review checklist is available here. Let me know if you have questions.
