### 2018-07-16 #821 @khaors
@leeper In the Installation Instructions, the list of dependencies should be clearly stated. Is this referring  to the README file in addition to the DESCRIPTION file?
