### 2018-07-16 robwschlegel/heatwaveR#7 @khaors
JOSS Review
Hi Rob. It has been a pleasure to review your R package. Really Good work.  During the review I have found the following minor issues:

- The release version 0.3.0 does not match the GitHub release 0.2.7. Can you check this?
- I didn't see any section in the README file with guidelines to third parties wishing to contribute to your R package. It would be great if you include a badge to invite people to contribute to the development of heatwaveR.
- In the README file, you include a really great description of the functionality of heatwaveR and how to use it in a simple example. However, it would be great if these items are included in a separate vignette titled "Introduction to heatwaveR package" or something like that. My point here is that this information about the package and the simple example are available only in the GitHub repository but not inside R when the package is installed. This will free some space in your README to include the contribution badge and to highlight your contact information to report bugs.
- When the R package is installed, the title of the complex_clims vignette is not visible in the _Vignettes and other documentation_ section in the package documentation. Can you check this?
