### 2018-07-16 #821 @leeper
@khaors DESCRIPTION file is fine, unless there are non-CRAN, non-R, or otherwise atypical dependencies.
