### 2018-07-19 #821 @khaors
@leeper I have finished the review. Sorry for the delay. Checking the vignettes took a while but the results are ok. I opened an issue in the heatwaveR repository with the my comments ( you can check this under the tag JOSS review).
