### 2018-07-19 #821 @leeper
@khaors Excellent. Thanks for your careful review!

@robwschlegel Please address the issues raised at: https://github.com/robwschlegel/heatwaveR/issues/7 and let me know when you're ready for me to take another look.
