### 2018-07-19 #821 @robwschlegel
Hello,
Thank you very much for reviewing the package.
I've read over the comments and am happy to make the changes. I'm
particularly glad that it was recommended to cut down the length of the
README and move that information to separate vignettes.
I've just returned now from a conference and have some work to catch up on
but will make the corrections as soon as I can. Hopefully within a week.
All the best,
-Robert

On Thu, 19 Jul 2018, 12:42 Thomas J. Leeper <notifications@github.com>
wrote:

> @khaors <https://github.com/khaors> Excellent. Thanks for your careful
> review!
>
> @robwschlegel <https://github.com/robwschlegel> Please address the issues
> raised at: robwschlegel/heatwaveR#7
> <https://github.com/robwschlegel/heatwaveR/issues/7> and let me know when
> you're ready for me to take another look.
>
