### 2018-07-23 d5edfdc
Adding vignettes back into package
