+++ vignettes/gridded_event_detection.Rmd

@@ -1,0 +1,197 @@
+---
+title: "Extreme event detection in gridded data"
+author: "AJ Smit"
+date: "`r Sys.Date()`"
+description: "This vignette covers the steps one may take to load netCDF data and analyse each pixel as separate time series for the occurrence of extreme events."
+output: 
+  rmarkdown::html_vignette:
+    fig_caption: yes
+vignette: >
+  %\VignetteIndexEntry{Extreme event detection in gridded data}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+```{r global_options, include = FALSE}
+knitr::opts_chunk$set(fig.width = 4, fig.align = 'center',
+                      echo = FALSE, warning = FALSE, 
+                      message = FALSE, tidy = FALSE)
+```
+
+```{r load-pkg, echo = FALSE, eval = FALSE}
+library(RmarineHeatWaves)
+library(ncdf4)
+library(plyr)
+library(dplyr)
+library(tibble)
+library(reshape2)
+library(lubridate)
+library(ggplot2)
+library(doMC); doMC::registerDoMC(cores = 4)
+```
+
+## Overview
+
+Because of human-induced climate change, we anticipate that extreme events would occur more frequently and that they would become greater in intensity. Here I address this question by using some gridded SST data, which is the only way that we can assess if this trend is unfolding across large ocean regions. Using the gridded [0.25 degree Reynolds OISST](http://dx.doi.org/10.1175/2007JCLI1824.1), I will find extreme SST events around South Africa data by applying the `detect()` function pixel-by-pixel. After detecting the events, I will fit a generalised linear model (GLM) to calculate rates of change in some marine heat wave metrics, and then plot the estimated trend. 
+
+## Reading in the netCDF
+
+These instructions assume we start with a netCDF file that contains data for only the region of interest in our study, with all of the daily data in one file. If you start from scratch, it might make sense to rather do the subsetting elsewhere and process the daily netCDF files. If that is the case, see [this vignette](https://robwschlegel.github.io/heatwaveR/articles/OISST_preparation.html).
+
+First, I read in the 0.25 degree [OISST netCDF file](https://www.ncei.noaa.gov/thredds/blended-global/oisst-catalog.html). I save the only variable, **sst**, as a multidimensional array (longitude, latitude and date). To do this, I use the **ncdf4** package. I then construct a list with names and values for each dimension of the SST array. The example data used here cover 93 longitude and 43 latitude steps (i.e. 3999 pixels), and cover 12797 days (1981 to 2016). The names for the three dimensions are simply a list of longitudes, latitudes and dates, which I get directly from the netCDF file. Note that not all netCDF files are created equally, and sometimes vectors of longitudes, latitudes and dates need to be created based on the start values for lat/lon/date in the netCDF, and the 'steps' between sequential lats/lons/dates. To determine out how to approach reading in the netCDF file, use the `print()` function to scrutinise the content of the netCDF.
+
+```{r read-netcdf, echo = TRUE, eval = FALSE}
+nc <- nc_open("/Users/ajsmit/spatial/OISSTv2/daily/OISST-V2-AVHRR_agg.nc")
+# print(nc)
[…]
