+++ vignettes/complex_clims.Rmd
@@ -1,0 +1,187 @@
+---
+title: "Complex Climatologies"
+author: "Robert W Schlegel"
+date: "`r Sys.Date()`"
+description: "This vignette walks the user through a case study on the application of a more complex climatology calculation."
+output: 
+  rmarkdown::html_vignette:
+    fig_caption: yes
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+bibliography: bibliography.bib
+---
+
+```{r global_options, include = FALSE}
+knitr::opts_chunk$set(fig.width = 8, fig.height = 3, fig.align = 'centre',
+                      echo = TRUE, warning = FALSE, message = FALSE,
+                      eval = TRUE, tidy = FALSE)
+```
+
+## Overview
+
+One of the long term goals of __`heatwaveR`__ is the inclusion of many different methods for the creation of climatologies for the use of detecting heatwaves and cold-spells in time series data. To this end we have made a very large change in the event detection pipeline, which is why we moved from the __`RmarineHeatWaves`__ package to __`heatwaveR`__. This change was the inclusion of the `ts2clm()` function and the removal of climatology generating found in `RmarineHeatWaves::detect()` in favour of `detect_event()`, which does not calculate climatologies. In this way we have allowed for the introduction of a multitude of more complex climatology calculation and event detection methods. It is our overarching goal to provide one package that allows climate scientists to calculate these events in both the atmosphere and oceans. But rather than talking about it, let's walk through a case study on how this shift in the main pipeline of this package can be used for diverse applications.
+
+Brought to our attention by Mr. Haouari from the IHFR institute of meteorology in Algeria was the concept of using a flat 25$^\circ$C `tMin` bottom boundary to screen out events calculated from `tMax` with the standard 90th percentile upper threshold. As the authors of the __`heatwaveR`__ package are admittedly marine oriented, we tend to work with daily time series that have only one mean value per day. The use of `tMin` and `tMax` is therefore not accommodated explicitly in the arguments that one gives to the `ts2clm()` and `detect_event()` functions, but that does not mean that one cannot do so. Below we will work through the steps one would take to calculate (atmospheric) heatwaves, as per the definition for them laid out in @Perkins2013, but excluding the calculation of EHF, and with the additional step proposed by Mr. Haouari. In the interest of reproducibility, we will be creating a `tMin` and `tMax` time series from the `sst_WA` data that is installed with __`heatwaveR`__. This is not technically correct to do, but will allow us to illustrate the methodology.
+
+## The methodology
+
+In the following sub-sections we will walk through the step-by-step approach needed to calculate atmospheric heatwaves using a 90th percentile threshold created from the `tMax` time series for a location, and then filter the events based on the `tMin` series also needing to exceed 25$^\circ$C on the same days. We will finish by showing how to then convert these results back into a format that `event_line()` and `lolli_plot` like so that one may still use these convenient functions to visualise the results.
+
+### Data prep
+
+The first step with any analysis in R should be the loading of the packages to be used.
+
+```{r load-libs}
+library(tidyverse)
+library(heatwaveR)
+```
+
+With our libraries loaded, we will now go about creating artificial `tMin` and `tMax` time series from `sst_WA`. Again, please note that this is not actually something that one should do. We only do so here to illustrate how one would go about doing this. Real `tMin` and `tMax` time series should be used when one is executing this methodology for proper research.
+
+```{r data-prep}
+# Create tMin time series
+tMin <- sst_WA %>% 
+  mutate(temp = temp - 1)
+
[…]

[…]
