### 2018-07-23 785f164
Edits suggested through JOSS review
