+++ vignettes/detection_and_visualisation.Rmd
@@ -1,0 +1,161 @@
+---
+title: "Detection and Visualisation"
+author: "AJ Smit and Robert W Schlegel"
+date: "`r Sys.Date()`"
+description: "This vignette demonstrates the basic use of the heatwaveR package for the detection and visualisation of events."
+output: 
+  rmarkdown::html_vignette:
+    fig_caption: yes
+vignette: >
+  %\VignetteIndexEntry{Vignette Title}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+bibliography: bibliography.bib
+---
+
+```{r global_options, include = FALSE}
+knitr::opts_chunk$set(fig.width = 8, fig.height = 3, fig.align = 'centre',
+                      echo = TRUE, warning = FALSE, message = FALSE,
+                      eval = TRUE, tidy = FALSE)
+```
+
+## Overview
+
+The `detect_event()` function is the package's core function. Here is the `detect_event()` function applied to the Western Australian test data, which are also discussed by Hobday et al. (2016):
+
+```{r detect-example1}
+# Load libraries
+library(heatwaveR)
+library(dplyr)
+
+# Detect the events in a time series
+ts <- ts2clm(sst_WA, climatologyPeriod = c("1983-01-01", "2012-12-31"))
+mhw <- detect_event(ts)
+
+# View just a few metrics
+mhw$event %>% 
+  dplyr::ungroup() %>%
+  dplyr::select(event_no, duration, date_start, date_peak, intensity_mean, intensity_max, intensity_cumulative) %>% 
+  dplyr::arrange(-intensity_cumulative) %>% 
+  head(5)
+```
+
+The corresponding `event_line()` and `lolli_plot()`, which represent the massive Western Australian heatwave of 2011, look like this:
+
+```{r fig-example1, echo = TRUE, eval = TRUE}
+event_line(mhw, spread = 100, metric = "intensity_cumulative", start_date = "2010-11-01", end_date = "2011-06-30")
+```
+
+<!-- ![](vignettes/fig-example1-1.png) -->
+
+```{r fig-example2, echo = TRUE, eval = TRUE}
+lolli_plot(mhw)
+```
+
+<!-- ![](vignettes/fig-example2-1.png) -->
+
+The `event_line()` and `lolli_plot()` functions were designed to work directly on one of the dataframes in the list returned by `detect_event()`. If more control over the figures is required, it may be useful to create them in **`ggplot2`** by stacking `geoms`. We specifically created two new **`ggplot2`** `geoms` to reproduce the functionality of `event_line()` and `lolli_plot()`. These functions are more general in their functionality and can be used outside of the **`heatwaveR`** package, too. To apply them to MHWs and MCSs first requires that we access the `clim` or `event` dataframes within the list that is produced by `detect_event()`. Here is how:
+
+```{r fig-example3, echo = TRUE, eval = TRUE}
+# select the region of the time series of interest
+mhw2 <- mhw$clim %>% 
+  slice(10580:10690)
+
+ggplot(mhw2, aes(x = t, y = temp, y2 = thresh)) +
+  geom_flame() +
+  geom_text(aes(x = as.Date("2011-02-26"), y = 25.8, label = "the Destroyer\nof Kelps"))
+
+ggplot(mhw$event, aes(x = date_start, y = intensity_max)) +
+  geom_lolli(colour = "salmon", colour_n = "red", n = 3) +
+  geom_text(colour = "black", aes(x = as.Date("2006-08-01"), y = 5,
[…]
