+++ README.Rmd
@@ -126,287 +126,21 @@ The events are summarised using a range of event metrics:
 
 Note that `rate_onset` and `rate_decline` will return `NA` when the event begins/ends on the first/last day of the time series. This may be particularly evident when the function is applied to large gridded data sets. Although the other metrics do not contain any errors and provide sensible values, please take this into account in the interpretation of the output.
 
-## Examples of detection and graphing
+## Vignettes
 
-The `detect_event()` function is the package's core function. Here is the `detect_event()` function applied to the Western Australian test data, which are also discussed by Hobday et al. (2016):
+For more detailed explanations on the use of the __`heatwaveR`__ package please follow the links below:
+  * For a basic introduction to the [detection and visualisation](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html) of events.
+  * For an explanation on the use of the [exceedence](https://robwschlegel.github.io/heatwaveR/articles/exceedence.html) function.
+  * For a walkthrough on the calculation and visualisation of [event categories](https://robwschlegel.github.io/heatwaveR/articles/event_categories.html)
+  * For examples on the calculation of [complex climatologies](https://robwschlegel.github.io/heatwaveR/articles/complex_clims.html).
+  * For a demonstration on how to [download and prepare OISST data](https://robwschlegel.github.io/heatwaveR/articles/OISST_preparation.html).
+  * Which may then have the `detect_event()` function [applied to the 3D gridded data](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html), and then fit a GLM and plot the results.
 
-```{r detect-example1}
-library(heatwaveR); library(dplyr); library(ggplot2)
-ts <- ts2clm(sst_WA, climatologyPeriod = c("1983-01-01", "2012-12-31"))
-mhw <- detect_event(ts)
-mhw$event %>% 
-  dplyr::ungroup() %>%
-  dplyr::select(event_no, duration, date_start, date_peak, intensity_mean, intensity_max, intensity_cumulative) %>% 
-  dplyr::arrange(-intensity_cumulative) %>% 
-  head(5)
-```
+## Contributing to __`heatwaveR`__ 
 
-The corresponding `event_line()` and `lolli_plot()`, which represent the massive Western Australian heatwave of 2011, look like this:
+To contribute to the package please follow the guidelines [here](CONTRIBUTING.md).
 
-```{r fig-example1, echo = TRUE, eval = FALSE}
-event_line(mhw, spread = 100, metric = "intensity_cumulative", start_date = "2010-11-01", end_date = "2011-06-30")
-```
-
-![](vignettes/fig-example1-1.png)
-
-```{r fig-example2, echo = TRUE, eval = FALSE}
-lolli_plot(mhw)
-```
-
-![](vignettes/fig-example2-1.png)
-
[…]
