### 2018-07-23 49e57d8
Further tweaks
