### 2018-07-23 1b5314b
Delete gridded_event_detection.Rmd
