### 2018-07-23 f12bda2
tweak to links in README
