+++ README.Rmd
@@ -131,11 +131,11 @@ Note that `rate_onset` and `rate_decline` will return `NA` when the event begins
 For detailed explanations and walkthroughs on the use of the __`heatwaveR`__ package please click on the articles tab above, or follow the links below:
 
   * For a basic introduction to the [detection and visualisation](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html) of events.
-  * For an explanation on the use of the [exceedence](https://robwschlegel.github.io/heatwaveR/articles/exceedence.html) function.
+  * For an explanation on the use of the [exceedence](https://robwschlegel.github.io/heatwaveR/articles/exceedance.html) function.
   * For a walkthrough on the calculation and visualisation of [event categories](https://robwschlegel.github.io/heatwaveR/articles/event_categories.html).
   * For examples on the calculation of [complex climatologies](https://robwschlegel.github.io/heatwaveR/articles/complex_clims.html).
   * For a demonstration on how to [download and prepare OISST data](https://robwschlegel.github.io/heatwaveR/articles/OISST_preparation.html).
-  * Which may then have the `detect_event()` function applied to the [3D gridded data](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html), and then fit a GLM and plot the results.
+  * Which may then have the `detect_event()` function applied to the [gridded data](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html), and then fit a GLM and plot the results.
 
 ## Contributing to __`heatwaveR`__ 
 
