+++ README.md
@@ -160,345 +160,24 @@ The events are summarised using a range of event metrics:
 
 Note that `rate_onset` and `rate_decline` will return `NA` when the event begins/ends on the first/last day of the time series. This may be particularly evident when the function is applied to large gridded data sets. Although the other metrics do not contain any errors and provide sensible values, please take this into account in the interpretation of the output.
 
-Examples of detection and graphing
-----------------------------------
+Vignettes
+---------
 
-The `detect_event()` function is the package’s core function. Here is the `detect_event()` function applied to the Western Australian test data, which are also discussed by Hobday et al. (2016):
+For detailed explanations and walkthroughs on the use of the **`heatwaveR`** package please click on the articles tab above, or follow the links below:
 
-``` r
-library(heatwaveR); library(dplyr); library(ggplot2)
-ts <- ts2clm(sst_WA, climatologyPeriod = c("1983-01-01", "2012-12-31"))
-mhw <- detect_event(ts)
-mhw$event %>% 
-  dplyr::ungroup() %>%
-  dplyr::select(event_no, duration, date_start, date_peak, intensity_mean, intensity_max, intensity_cumulative) %>% 
-  dplyr::arrange(-intensity_cumulative) %>% 
-  head(5)
-#> # A tibble: 5 x 7
-#>   event_no duration date_start date_peak  intensity_mean intensity_max
-#>      <int>    <int> <date>     <date>              <dbl>         <dbl>
-#> 1       22       95 1999-05-13 1999-05-22           2.50          3.60
-#> 2       42       60 2011-02-06 2011-02-28           3.21          6.51
-#> 3       49       47 2012-01-11 2012-01-27           2.23          3.30
-#> 4       50       46 2012-03-01 2012-04-10           1.99          2.96
-#> 5       41       40 2010-12-24 2011-01-28           2.16          3.27
-#> # ... with 1 more variable: intensity_cumulative <dbl>
-```
+-   For a basic introduction to the [detection and visualisation](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html) of events.
+-   For an explanation on the use of the [exceedance](https://robwschlegel.github.io/heatwaveR/articles/exceedance.html) function.
+-   For a walkthrough on the calculation and visualisation of [event categories](https://robwschlegel.github.io/heatwaveR/articles/event_categories.html).
+-   For examples on the calculation of [complex climatologies](https://robwschlegel.github.io/heatwaveR/articles/complex_clims.html).
+-   For a demonstration on how to [download and prepare OISST data](https://robwschlegel.github.io/heatwaveR/articles/OISST_preparation.html).
+-   Which may then have the `detect_event()` function applied to the [gridded data](https://robwschlegel.github.io/heatwaveR/articles/gridded_event_detection.html), and then fit a GLM and plot the results.
 
-The corresponding `event_line()` and `lolli_plot()`, which represent the massive Western Australian heatwave of 2011, look like this:
+Contributing to **`heatwaveR`**
+-------------------------------
 
-``` r
-event_line(mhw, spread = 100, metric = "intensity_cumulative", start_date = "2010-11-01", end_date = "2011-06-30")
-```
+To contribute to the package please follow the guidelines [here](ttps://robwschlegel.github.io/heatwaveR/CONTRIBUTING.html).
 
[…]
