+++ README.Rmd
@@ -139,7 +139,7 @@ For detailed explanations and walkthroughs on the use of the __`heatwaveR`__ pac
 
 ## Contributing to __`heatwaveR`__ 
 
-To contribute to the package please follow the guidelines [here](ttps://robwschlegel.github.io/heatwaveR/CONTRIBUTING.html).
+To contribute to the package please follow the guidelines [here](https://robwschlegel.github.io/heatwaveR/CONTRIBUTING.html).
 
 Please use this [link](https://github.com/robwschlegel/heatwaveR/issues) to report any bugs found.
 
