### 2018-07-25 4cb63d0
BUG FIX: corrected issue with `clim_calc_cpp` not being able to calculate   clims from baselines not beginning and ending on the Julian year by making   `clim_spread` plug the gaps beforehand with rowwise means.
