### 2018-07-25 6482a70
Double clim examples added to vignette and R file
