+++ DESCRIPTION
@@ -18,6 +18,7 @@ RoxygenNote: 6.0.1
 Depends: R (>= 3.0.2), data.table, ggplot2
 Suggests:
     tidyverse,
+    ggpubr,
     testthat,
     knitr,
     rmarkdown
