### 2018-07-25 robwschlegel/heatwaveR#7 @robwschlegel
Hello,
Apologies for the delay. I've gone over the suggested edits and have my responses itemised below:

- You must have checked the package just as the CRAN version was still being updated to 0.3.0. While this review process was happening my co-author fixed a couple of bugs and so now the version has gone up to 0.3.3. This is the version presently on GitHub. I submitted the source package of 0.3.3 to CRAN today and so it should be reflected there within a week or so.
- In the README file I have now added a section at the end that provides a link with instruction for how to contribute to the package.
- I've taken much of the information on calculating heatwaves and visualising them out of the README and added a few new vignettes that now contain that information.
- I assume that the complex_clims vignette didn't show up in the package because it was not in the version on CRAN when the review occurred. Once the version on CRAN is updated to 0.3.3 this should be resolved.

All the best,
-Robert
