### 2018-07-25 #821 @robwschlegel
I've addressed the issues at robwschlegel/heatwaveR#7
