### 2018-07-25 #821 @leeper
Great, thank you @robwschlegel. @khaors Can you take a look and see if there's anything else that should be addressed?
