### 2018-07-26 0491e1d
Minor edits.
