### 2018-07-31 #821 @khaors
@leeper I checked and the suggestions have been addressed and the heatwaveR package has been modified in accordance.
