### 2018-07-31 #821 @leeper
@robwschlegel The review process is now complete. To finalize your submission and accept your paper in JOSS, we need you to deposit a copy of your software repository (including any revisions made during the JOSS review process) with a data-archiving service. To do so:

 3. Post a comment here to @leeper with the DOI for the release.

Let me know if you have any questions about the process.
