+++ README.Rmd
@@ -24,6 +24,7 @@ knitr::opts_chunk$set(
 [![CRAN_Status_Badge](http://www.r-pkg.org/badges/version/heatwaveR)](https://cran.r-project.org/package=heatwaveR)
 [![Travis build status](https://travis-ci.org/robwschlegel/heatwaveR.svg?branch=master)](https://travis-ci.org/robwschlegel/heatwaveR)
 [![Coverage status](https://codecov.io/gh/robwschlegel/heatwaveR/branch/master/graph/badge.svg)](https://codecov.io/github/robwschlegel/heatwaveR?branch=master)
+[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.1324309.svg)](https://doi.org/10.5281/zenodo.1324309)
 ![](https://cranlogs.r-pkg.org/badges/grand-total/heatwaveR)
 
 The __`heatwaveR`__ package is a project-wide update to the [__`RmarineHeatWaves`__](https://github.com/ajsmit/RmarineHeatWaves) package,
