+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: heatwaveR
 Version: 0.3.3
-Date: 2018-07-25
+Date: 2018-07-31
 Title: Detect Heatwaves and Cold-Spells
 Description: The different methods of defining and detecting extreme events, known as heatwaves or cold-spells in both air and water temperature data are encompassed within this package. These detection algorithms may be used on non-temperature data as well however, this is not catered for explicitly here as no use of this technique in the literature currently exists.
 Type: Package
